+++ firebase-vscode/webviews/globals/ux-text.ts
@@ -1,0 +1,34 @@
+export const TEXT = {
+  INIT_HOSTING_BUTTON: "Host your Web App",
+
+  INIT_HOSTING_DESCRIPTION: "Deploy your app with Firebase Hosting"
+    + ", a high-performance static web host backed by a global CDN",
+
+  INIT_HOSTING_PROGRESS: "Initializing...",
+
+  LOGIN_PROGRESS: "Checking login",
+
+  MONOSPACE_LOGGED_IN: "Using default credentials",
+
+  MONOSPACE_LOGIN_SELECTION_ITEM: "Default credentials",
+
+  VSCE_SERVICE_ACCOUNT_LOGGED_IN: "Logged in with service account",
+
+  VSCE_SERVICE_ACCOUNT_SELECTION_ITEM: "Service account",
+
+  MONOSPACE_LOGIN_FAIL: "Unable to find default credentials",
+
+  GOOGLE_SIGN_IN: "Sign in with Google",
+
+  ADDITIONAL_USER_SIGN_IN: "Sign in another user...",
+
+  SHOW_SERVICE_ACCOUNT: "Show service account email",
+
+  CONSOLE_LINK_DESCRIPTION: "Open in Firebase console",
+
+  CONNECT_FIREBASE_PROJECT: "Connect a Firebase project",
+
+  DEPLOYING_IN_PROGRESS: "Deploying...",
+
+  DEPLOYING_PROGRESS_FRAMEWORK: "Deploying... this may take a few minutes."
+};