+++ src/extensions/paramHelper.ts
@@ -7,7 +7,6 @@ import { logger } from "../logger";
 import { ExtensionInstance, ExtensionSpec, Param } from "./types";
 import { getFirebaseProjectParams, substituteParams } from "./extensionsHelper";
 import * as askUserForParam from "./askUserForParam";
-import { track } from "../track";
 import * as env from "../functions/env";
 import { cloneDeep } from "../utils";
 
@@ -95,7 +94,6 @@ export async function getParams(args: {
   instanceId: string;
   paramSpecs: Param[];
   nonInteractive?: boolean;
-  paramsEnvPath?: string;
   reconfiguring?: boolean;
 }): Promise<Record<string, ParamBindingOptions>> {
   let params: Record<string, ParamBindingOptions>;
@@ -111,8 +109,6 @@ export async function getParams(args: {
       reconfiguring: !!args.reconfiguring,
     });
   }
-  const paramNames = Object.keys(params);
-  void track("Extension Params", paramNames.length ? "Not Present" : "Present", paramNames.length);
   return params;
 }
 
@@ -121,7 +117,6 @@ export async function getParamsForUpdate(args: {
   newSpec: ExtensionSpec;
   currentParams: { [option: string]: string };
   projectId?: string;
-  paramsEnvPath?: string;
   nonInteractive?: boolean;
   instanceId: string;
 }): Promise<Record<string, ParamBindingOptions>> {
@@ -137,8 +132,6 @@ export async function getParamsForUpdate(args: {
       instanceId: args.instanceId,
     });
   }
-  const paramNames = Object.keys(params);
-  void track("Extension Params", paramNames.length ? "Not Present" : "Present", paramNames.length);
   return params;
 }