+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- firebase functions:secrets:\* ensure the secretmanager API is enabled (#5918)
+- Fix bug where functions:secrets:\* family of commands did not work when Firebase CLI is authenticated via GOOGLE_APPLICATION_CREDENTIALS (#6190)